Add new mockup for DNS resolution session
Essentially the new features I'm considering are transient roles and
invite scopes.

A transient role is a role which isn't populated at the start of the
conversation, by either the role |-> actor mapping or some config file.

An invite scope is of the following form:

X invites Y {
  // Y is valid in this scope, but not outside of it
}

Early idea. Might revise.