Refine failure model for sending messages
Previously failures with sending could not be detected.

Now, any send failures will be explicitly detected and an exception
thrown. This means in case of the Role |-> Endpoint mapping being stale
due to supervision resets, or monitor failure, the client code can
either Let It Fail and terminate, or handle the exception more
gracefully.