Revamp invitation / init system, fix become
Quite a few big changes today.

In particular, conversation initiation is now asynchronous. This
actually makes the API a lot nicer and prevents a class of deadlocks.
Actors can also either accept or decline the invitation.

Before, become was broken as it wasn't possible to disambiguate between
different instances of a protocol-role conversation. Now, I've added a
feature whereby conversation instances can be registered, allowing them
to be switched to later on. The rationale behind this is that I think
that the number of situations in which "become" will be necessary is
quite small -- in particular, it's likely not going to be necessary to
switch to a worker thread, but more likely a single distinguished thread
taking part in a long-running conversation.