Large re-architecting regarding monitors
Monitors are now defined at a conversation level, as opposed to an actor
one. There are a few reasons for this change, but the main one is that
we can migrate monitors way more easily in the case of actor failure.

This marks an end to actor_monitor, which is now replaced by actor_proxy
(which keeps track of session-related state for actors, but *not*
monitors), and role_monitor, which defines a monitor for a role in a
session.